+++ CRM/src/main/java/com/sc/mapper/SysBranchMapper.java
@@ -2,7 +2,6 @@ package com.sc.mapper;
 
 import com.sc.bean.SysBranch;
 import com.sc.bean.SysBranchExample;
-import java.math.BigDecimal;
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
@@ -11,7 +10,7 @@ public interface SysBranchMapper {
 
     int deleteByExample(SysBranchExample example);
 
-    int deleteByPrimaryKey(BigDecimal secId);
+    int deleteByPrimaryKey(Long secId);
 
     int insert(SysBranch record);
 
@@ -19,7 +18,7 @@ public interface SysBranchMapper {
 
     List<SysBranch> selectByExample(SysBranchExample example);
 
-    SysBranch selectByPrimaryKey(BigDecimal secId);
+    SysBranch selectByPrimaryKey(Long secId);
 
     int updateByExampleSelective(@Param("record") SysBranch record, @Param("example") SysBranchExample example);
